Fix #13: merge ArcCheck and MapCheck parsers
